[WIP] pytest koolaid
test_pyteststuff.py is a proof of concept of using indirect
parameterized fixtures to generate all the tests we want, and sharing
those fixtures across all the tests that need them.

Now all that remains is to clean up Probe/ProbeResult and port all the
tests.

For bonus points, we can probably trade in run_webserver for a fixture
and avoid having to run that once(-ish) per test.