Long-running probe w/ stdio communications
Also a tocluster test which uses the stdio communications to direct the
probe to make some HTTP requests and report the results.  Combined with
the existing environment variable tests, this represents a near-complete
port of the tocluster test suite.  Some ``--also-proxy`` tests remain to
be ported.  Beyond that, some tests which appear to be for to-cluster
behavior but which are actually for argument parsing need to be
rewritten to avoid running Telepresence.