Sometimes kubectl breaks
That is, it exits with a success code and also no output whatsoever.
This must be bogus behavior because the thing we run always produces
some output.  If we don't get any output from it, kubectl must have
broken.  Yet kubectl hasn't exited with an error code.

So... try to handle this case with the empty-retry logic and avoid
assuming anything about res if it's empty.