Makefile: (almost) Exclusively refer $(KUBECONFIG) to refer to cluster.knaut
This allows the caller to override the value by passing a `KUBECONFIG=…`
argument to `make`.  This is an important property for running `make check`
inside of `testbench`, so that testbench can provide the Kubernaut file.

The exception is that we still hard-code the value in

    clean: cluster.knaut.clean

so that `make clean` doesn't remove the user's Kubernaut file.