go-mod.mk: Enable parallel build, drop ".NOTPARALLEL:"
I had written ".NOTPARALLEL:" because in Go 1.11, it is not safe for `go`
commands that write to the module cache (${GOPATH}/pkg/mod/) to be invoked
concurrently (this is expected to be fixed in Go 1.12 in February 2019).
Instead of disabling parallel Make, have every Make target that might
write to the module cache depend on "go-get", so that the cache should be
full, and nothing will have to write to it.

The only place that this require a change is making "go-fmt" depend on
"go-get".