pkg/k8s: Rename resourceInfo→ResourceType, Client.resolve→Client.ResolveResourceType, improve docs
"resource type" is a more descriptive name than "resource info" (which
could be taken to mean "information about an instance of a resource")

"resolve resource type" is a more descriptive name than "resolve"; it
tells you *what* it resolves.

I had to make ResourceType public in order to make ResolveResourceType
public.  I wanted to make ResolveResourceType public because it's largely
(but not entirely, as my improved documentation points out) a UI concern;
callers should be able to call it directly for the UI.