pkg/dtest/testprocess: Don't have .MakeSudo trigger everything running as root
Before, if *any* of the commands wanted to be `sudo`, then it sent *all*
of the commands through sudo, and then `sudo -u $SUDO_USER` back to the
regular user.  Also, it makes the regular test code run as root.

Skip that step.

Also, while we're at it, get rid of the `sub` singleton; just have its
members be global variables.

Of course, the test code in cmd/teleproxy need to run as root, so have it
call dtest.Sudo() explicitly now.