pkg/kubeapply: Go back to having timeout be per-phase
After discussing this with @rhs a bit on GitHub, it might (IMO) make more
sense to the end-user to set the timeout in-total instead of per-phase, but
it is much harder to have sane defaults in-total than per-phase.

But I found it really surprising when I read the code and discovered that
the timeout was per-phase.  So update some variable names and --help text
to hopefully make that less surprising to people in the future.