.github/pull_request_template.md: Add a checkbox reminding you to update CHANGELOG
The intent has been that CHANGELOG.md be continually up-to-date, so that
when we want to cut a release, all that needs to happen in CHANGELOG.md
is to set the release date.

But that hasn't been happening; PRs that change things haven't touched
CHANGELOG.md.  Because of this, yesterday (2019-09-19), both Abhay and I
agreed that we should do a release, but neither of us really wanted to
interrupt what we were doing to go through the git log to see what
changed.

I noted that in apro.git, adding a checkbox to the PR template reminding
you to update the CHANGELOG in your PR has been super-helpful in making
sure that the CHANGELOG is always up-to-date.  We agreed that adding the
same thing to teleproxy.git would be a good idea.