Make Kubernetes and Consul watch channels buffered
Co-authored-by: Rafael Schloming <rhs@datawire.io>
Co-authored-by: Flynn <flynn@datawire.io>

In case of rapid additions and deletions of the same resource
makes watt become unresponsive. This broadly happens because watt
creates and deletes watches of the same name in rapid succession,
and after some time watt reaches a place where a watcher is trying
to shut down but it blocks on an aggregator notify event.

This commit makes watch channels buffered which gives the code
enough leeway to process rapid changes to the cluster.