fix: calls /users endpoint to determine if oc should be used
Previously called `/apis` endpoint which lists all resources in the
cluster can be secured making it impossible to determine if the actual
cluster is Openshift. In such a case Telepresence rolls back to
`kubectl` and assumes `Deployment` is used rather than `DeploymentConfig`
which is commonly used in Openshift.

Proposed solution calls Openshift-specific endpoint and based on the
response decides which cli tool to use.

Fixes #1139