Update __init__.py
Added verification if parameter --Deployment or -d has been provided. Without that verification an error is printed that the deploymentconfig NONE can not be found if the user has installed the oc tool. Furthermore added code in _dc_exists to use only the DeploymentConfig name and not the container name.