Ensure context is passed with all prelim_command
When a user is specifying a context, it may mean he is working outside the default context. If this is the case and the default context is not using the same versions that the context he will be working in, then the way this is currently written is checking the wrong versions (and failing, if the default context is failing).

This set's the context immediately after finding which `prelim_command` should be used. Thus, all subsequent checks will be checking the correct context versions.

Use double quotes to match formatting

Fix formatting

Add changelog entry

Fixes: #1190

Fix Popen arguments for prelim_command with context

Fix Incompatible types in assignment (expression has type "List[Any]", variable has type "str")