Fix pod CIDR detection when falling back to using pod IPs
If the pod CIDR can not be determined by looking at node
configuration, an attempt will be made to collect pod ips and find a
CIDR that covers them. Previously only pods in the current namespace
were collected which may not result in an accurate CIDR. This change
ensures that the IPs of pods from all namespaces are considered.