Switch to creating deployment with `kubectl create`
As of kube 1.18, `kubectl run` is no longer capable of creating a
deployment. This commit switches to supplying deployment yaml to
`kubectl create -f` to ensure compatibility with all supported
versions of kubectl. If `--expose` is specified, `kubectl expose` will
be used.

Note that the yaml specifies `apiVersion: apps/v1`, which is
compatible with kube >= 1.9.