Migrate to gRPC client-server architecture
This commit replaces the current JSON-based communication between the
`edgectl` client and daemon processes with a gRPC based API. The
API description and the files generated from it lives in the package
`internal/pkg/edgectl/rpc`.

The commit also separates the files belonging to the client and daemon
into two corresponding packages. The client package contains all code
pertaining to CLI and presentation details and daemon no longer produces
anything directly intended for the user (aside from logging to the
logfile).

Closes #1571