Don't consider failure to connect to traffic manager an error
This commit reverts the behavior of treating a traffic manager connect
failure as an error to just report "Connected to context" with info
that intercept is unavailable and then exit with zero.

The usefulness of this behavior is questionable but since treating it
as an error is a behavioral change, it can wait until we make more
such changes.