Create a service preview connector that can run as the user.
The edgectl daemon process must run as root in order to establish
network overrides but since all other tasks that it performs can run as
the invoking user, this commit moves them to a new background process
called a `connector`. The `connector` is started when the user issues
the `edgectl connect` command and terminates when the user issues a
`edgectl disconnect` or `edgectl quit` command.

The connector uses a new `daemon.Logger` method to send all log output
to the daemon which in turn will dispatch it to its own logger.

Some improvements were made in how the `daemon` terminates and ensures
termination of the connector so this commit closes #1841.

The status command no longer reports an error when the daemon isn't
running. Instead, it reports that status and exits with zero.

The errors and exit codes from the connect command are improved to
be more consistent. It's likely that this closes #1532.

The commit splits the former daemon Go package into the three distinct
packages `connector`, `daemon`, and `teleproxy`, each representing its
own executable. Care has been taken to export as little as possible from
these packages to make it easy to later facter them out into a separate
repository.

Closes #1334