Create an EnsurableState for the client side of the daemon.
Adds a `daemonState` which corresponds a possible connected client side
of the daemon. The `daemonState` implements the `EnsurableState` which
is needed later on for the `run` subcommand.

This commit also removes some printouts directed to `os.Stdout` in favor
of using the `io.Writer` provided by the command (this makes it easier
to capture output in tests later on).