Add edgectl run command
This commit adds `run -- <command>` to the set of available sub-commands
for `edgectl`. It is a convenience command that ensures that the daemon
is started, a connection has been made, and an intercept has been added
before the command executes. The command then cleans up the resources
that it acquired, leaving the state as it were before it was executed.