Move functions from file not built on Windows platforms
Some functions were not included when building for Windows which
resulted in compiler errors on the windows platform. This commit moves
them to more suitable locations.

The `runAsRoot()` and `start()` functions are moved from the file
`launch_daemon_unix.go` to `run.go`.
The `quiteLegacyDaemon()` function is moved to `daemon_state.go`