Add ip-tables API to the client daemon process.
This commit adds what's needed in order to alter IP-tables and DNS
search paths. The intention is that this service will remove the the
current https based service in the intercept package.

The addition resulted in some refactoring of how proto files are
organized and how the subsequent generation works. Now each proto ends
up in a package of its own.