Auto-install a published image of the traffic-manager.
Adds logic to the connector to discover and then update or create
the traffic manager service and deployment.

Some changes to the make system was required in order for this to work.
When building an image, the name of that image will be stored in a file
so that the image name can be used in a -ldflags -X to tell the binary
the name of the image to use. This means that an install or a build
will build the image unless that file exists. Once it does, the image
will not be rebuilt unless a `make image` is used explicitly.