Major refactoring change where supervisor is replaced with dgroup.
I'm sorry this couldn't be split into several commits but the way
supervisor and dgroup does things is fundamentally different. The dgroup
package is more focused and doesn't provide extras like dependencies
and ready state for go routines which caused quite a lot of reorg (for
the better, I'd say).

The go.mod currently refers a specific commit on the apro master branch
which of course is unacceptable long term.