Let initial Connect call provide status
Obtaining the status from the connector was a two step process which
was unnecessary. Either the connector isn't started, which means that
there's no session, or it is running, which means that its status can
be retrieved by the client's connect call.

This commit modifies the gRPC sligthly so that connect() returns all
the status information and removes the explicit status() call.