Make installer update container probes
A container with probes that references a container port by a symbolic
name didn't work after the installer had modified that port and made
it anonymous, which it did because the name was instead given to the
port of the agent sidecar.

This commit changes this so that the port no longer becomes anonymous.
Instead, a name is generated to replace the original name. All probe
references to the original name are also replaced. The original name
is stored in an annotation to enable it to be restored by an installer.