Introduce manifest modifying actions that can be done and undone
This commit introduces the concept of actions that somehow modify
a manifest. Part of the modification is always to store the metadata
describing the action as an annotation in the manifest that it modifies.

This will allow the action to be recreated from the manifest and undone
but the actual API for undo is not part of this commit.