Rely on error default format instead of explicitly calling Error()
This is just a change of style. All calls where `%s` is used in
combination with `err.Error()` are replaced by `%v` and just `err`.
Also, all calls that will use the default string format directly
(such as `fmt.Fprintln()`, or `dlog.Error()`) are passed the error
directly instead of it's `Error()` string.