Remove redundant port-forward to the traffic-manager
The bridge used to have a life of its own in a separate process and
hence had no way of knowing what port the daemon used for port-forwards
to the traffic-manager. A consequence of that was that it established
a port-forward of it's own (from 8022 to 8022). With the bridge managed
by the same process as the traffic-manager client, the bridge can use
that port once the client is has started its port-forward.

This commit removes the redundant port-forward from the bridge and
ensures that the bridge is started after the traffic-manager start has
completed.