Let kates.Watch include Deployments and move it to k8scluster
Changes the query for the the `kates.Watch` so that it includes
`deployments` and moves it to the `k8sCluster`. The `k8sCluster` now
maintains a live set of `Services`, `Deployments`, and `Pods`. It is
also responsible for sending updates to the DNS service in the daemon.

The `bridge` is now only responsible for doing SSH port-forwards.

The live data in the `k8sCluster` is now also used by the installer so
that it no longer needs to do remote queries for existing services and
deployments.