Add upload-binary make target to upload a tagged telepresence binary
The upload-binary target will require that `awscli` is installed, that
correct aws credentials have been configured, that the repository is
tagged using a valid semantic version and that the workspace is clean.
With all those prerequisites fulfilled, it uploads the `telepresence`
binary to the S3 `datawire-static-files` bucket using a name that
contains the version, os, and architecture.