Add logic that selects alternate approach for DNS resolution
The logic uses DBus to check if `systemd-resolved` is present, and
if so, creates a TUN device configured with a DNS server. An
`nslookup` is performed to verify that the setup works (i.e. that
`systemd-resolved` is active rather than not just running). If it
is, then the local DNS-server will run without fallback and search
path. The setting of search-path will instead affect the DNS
service configured for the TUN network interface.

If `systemd-resolved` cannot be used, then the local DNS-service
is configured as before.