connector: install_actions.go: Have internally-typed functions on hideContainerPortAction
`hideContanerPortAction.Do()` and friends (by necessity of
implementing the `singleAction` interface) take an generic
`kates.Object` object as the argument, and type-assert it to a
`*kates.Deployment` internally.

I find that makes it hard to follow.  Instead, create an internal
`do()` method that takes a `*kates.Deployment`, and have `Do()` just
do the type-assertion and call `do()`.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>