.circleci: promote-to-stable needs to run for each platform
I'm not sure that's the best approach, but it's the current approach
(stable.txt has GOOS/GOARCH in the URL path), and I didn't realize that
when I wrote the CircleCI config the way I did.  If I think the approach
should be changed, I can advocate for that later.