Don't listen to Docker gateway unless it is visible as a network.
On WSL2 platforms, the Docker gateway is managed by the host operating
system (Windows) and not visible to Linux. This commit checks if that
is the case, and if so, makes no attempt to bind do the gateway since
there's no chance it's been subjected to the nat logic anyway.