daemon: Simplify the tree of how goroutines get named and started
There's a lot of this calls this, and launches this in the background...

No.  Stop it.  Just have the main `run()` entrypoint launch everything.
There's only one ordering dependency, and it's trivial to handle in run().
Two of them ended up being short-lived things that totally didn't need to
be in goroutines.

And give the goroutines more than 3 character names, names that actually
tell us what it's doing:

    NAT      → firewall-configurator
    PXY      → firewall-to-socks
    DNS      → server-dns
    CFG      → (removed)
    service  → server-grpc
    outbound → (removed)

Because of the web of this-starting-that, the goroutines now known as
"firewall-configurator" and "firewall-to-socks" might have previously
shown up as children of "DNS", depending on which platform you're on.

And, of course, add some comments as I untangle this.

"Ignore space change" might be useful for viewing the changes in
`service.go`.