Have kubectl port-forward assign the local ports
Acquiring a port by listening to `:0` and then closing the listener
may result in port conflicts due to the race condition that may occur
between the time when the listener is closed and the port is re-acquired
for its final purpose.

This commit alters the `portForwardAndThen()` function so that kubectl
can assign the local ports for us while doing the port-forward.