Prevent remote mounts from happening when there's nothing to mount.
Sets the APP_MOUNTS environment to the directory used as the prefix
for all mounts, but only if the app actually has mounts. The agent will
not start the ssh daemon unless it is set. The SshPort will be zero
unless the daemon starts, and the connector will pick up on that and
refrain from invoking the port-forward and sshfs.