Break config out of k8sCluster so that it can be compared between calls
The main purpose of this commit is to enable detection of Kubernetes
configuration changes that aren't just namespace differences. If a
different context is given, or a different kubeconfig even, then the
connector and daemon must be restarted.

The three kubectl shell-outs to get the correct configuration are
replaced by a call to `configFlags.ToRawKubeConfigLoader()` which
incidentally also provides the restconfig needed when doing discovery of
the server (which enabled the removal of yet another kubectl shell-out).