Watch namespaces dynamically and sync DNS with NAT
Modifies the k8sCluster so that it no longer watches all namespaces.
Instead, it watches a selection of namespaces returned by a watcher
of namespaces. The watchers reconfigure and updates the DNS and
IP-tables dynamically as namespaces come and go in the cluster.

The code is designed with further improvements in mind such as figuring
out what namespaces that are meaningful to watch when taking permissions
into account or simply let the user decide a subset that is meaningful
for a specific use-case.