Ensure that agent never receives intercepts for the wrong namespace.
A check was missing in the manager which made it send intercepts for a
specific agent name to all agents with that name irrespective of
namespace. The agent then tested if the intercept was intended for them.

This commit adds the misisng check to the manager and removes it from
the agent.