Simplify MacOS DNS-resolver and remove risk of race for local listener
I just discovered that the MacOS resolver files can have a port entry.
That opens up for a much simpler design where we can use a random port
on localhost instead of finding an IP where port 53 is available.

This commit also changes how the listener for the DNS service is
acquired so that it's acquired once instead of twice with the risk of
a race condition in between.