Make install_test not fail under some circumstances
The install_test used a fixed version number which sometimes could
make it use an older version of the traffic-manager. There was also
some difference in how `ko` was used that caused problems. This commit
changes the version number to include a the PID and removes the custom
`ko` build to instead call `make push-image`.