Fix logic so new case is interceptable
I noticed in doing this work that if a service has a targetPort that if
an int and the container has a containerPort + a name, the
intercept doesn't work. This is bc we update the service to use the name of the
containerPort while also using that name for the traffic-agent's port,
without changing the name of the containerPort.

I've fixed that here + included a test.