Make tail -F work without a "Permission denied" failure
Using `tail -F <file name>` enables the tailing of rotating log files.
Unfortunately, the daemon.log can be owned by root for a very short
while during the actual rotation and sometimes that results in a
"Permission denied" from `tail` and that further output stops.

This commit corrects the problem by creating the initial file with
the suffix ".tmp" and then changing the ownership of that file before
renaming it to its proper name (without ".tmp") so that `tail` never
gets the chance to attempt an open until the ownership is correct.