Add fake domain that can be used as varying search-path.
This commit addresses the intricate problem that Docker on MacOS (and
possibly on other platforms as well) overrides the the DNS search. It
uses the DNS-resolver and everything else works (including
NAME.NAMESPACE). It's just the DNS search that gets overridden. The
Docker dns-search can however be configured, either by passing options
to the invocation or by modifying the Docker config. Only problem is,
that the Telepresence search path changes dynamically. A fix for
that problem is to introduce a static fake domain (here named
"tel2-search"), which, when queried, will replace itself with the
dynamic search path.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>