go.mod: Drop now-pointless 'replace golang.org/x/sys' line
golang.org/x/sys commit 6fcdbc0 (unix: drop support for pre-1.12 direct
syscalls on darwin, 2020-08-25) dropped SYS_IOCTL, which we need.  So we
added a 'replace' line to our go.mod to pin to 6fcdbc0^.

However, the later commit c5abc1b (unix: revert zsysnum_darwin_*.go
deletion from CL 250437, 2020-10-28) reverted that, so we can drop the
'replace' as long as we depend on a new enough version.

And indeed, the commit referred to in the 'require' block is new enough;
it is f84b799 (2020-11-18).

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>