Uninstall once per deployment instead of per agent
So I noticed that I could only repro this error w/ my change
in the previous commit and only when I did uninstall --all-agents.

The reason for that is in the NAMED_AGENTS case it already picks
one agent per deployment.  But in the --all-namespaces case, we were
running into a problem where there could be more than one replica for
each deployment, and therefore multiple agents per deployment, so we
were removing the agent multiple times for deployments of more than
one replica which opened us up to race conditions with the deployment
and why we were seeing:
"deployments.apps "with-probes": the object has been modified".
This should fix it

Signed-off-by: Donny Yung <donaldyung@datawire.io>