Have tests only pass relevant information to make when building image
Something changes with go1.16 so that the environment in the executable
contains a lot more settings when it is executed with `go test`. These
environments were passed on to `make` and to the `ko` executable which
resulted in images being built for the host OS rather than the default
`linux/amd64` which in turn made all tests fail on a timeout when the
traffic-manager pod attempted to use the image and backed off.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>