Rename DeploymentInfo to WorkloadInfo
Since we are adding ReplicaSet support (and potentially future workloads
in the future e.g. Stateful sets), this commit changes DeploymentInfo to
WorkloadInfo and adds an additional field for the workload type.  I
think this encompasses all the places deploymentInfo was mentioned but
there might be some remnants in future commits.

Signed-off-by: Donny Yung <donaldyung@datawire.io>