Adds ReplicaSets
This is the bulk of the work for this ticket.  It shifts us from using
Deployments to using a more generic kates.Object which enables us to
support other workload types (e.g. ReplicaSets here) but can be further
extended in the future as we have more workload types (e.g.
StatefulSets)

Signed-off-by: Donny Yung <donaldyung@datawire.io>