Add service as intercept argument
Currently if there are multiple services that match the labels of the
workload, we use whichever one we find first.  This isn't really great
in the Argo CD case because there is a <service>-stable and
<service>-canary service that both have the same labels, so one
essentially is chosen at random.  Now users can specifiy the specific
service they want to intercept

Signed-off-by: Donny Yung <donaldyung@datawire.io>