Fix ReplicaSet intercept race condition
When refreshing the ReplicaSet pods, it was looking at all pods in the
namespace so if we had a pod that was terminating, the ki.findPod call
would error when it got to that pod. This has been fixed by only calling
findPod on pods that have the ReplicaSet name in their name.  This was
causing
TestTelepresence/TestB_Connected/TestG_SuccessfullyInterceptsReplicaSet
to be a little flakey.

Signed-off-by: Donny Yung <donaldyung@datawire.io>