Add RBAC Regression testing
I've updated our tests so that we use an RBAC user when we run a portion
of the pkg/client/cli/telepresence_test.go tests.  This will enable us
to catch when we make rbac changes (as long as we test our new features)
and ensure we keep the minimally necessary rbac file up-to-date.

Signed-off-by: Donny Yung <donaldyung@datawire.io>