Add license command
This adds support for a license command that given a license ID, will
communicate with system a to get the jwt formatted version of the
license, place it in a secret, and write that secret to a file on the
users machine.  This will be used for air-gapped environments as
non-air-gapped environments communicate with system a for their
licensing.

Signed-off-by: Donny Yung <donaldyung@datawire.io>