(from getambassador.io) Reorganize telepresence install docs
I've made two changes here:
1. Move the install old versions sections from install/upgrade to the
install page. We've had a few people ask about how to install older
versions and I think it's because upgrade (which is sorta the inverse of
what installing old versions is) doesn't seem like an obvious place to
work. So I think moving this to install will be clearer for users.

2. I've moved the migration chunk out of the upgrade file and into it's
own (see next commit). I think once we replace the docs on
telepresence.io with these docs and telepresence 2 is ~ the ~
telepresence moving forward, having a specific doc that is for
"migrating" seems important