Add TODO comment
As noted by Luke: Does this ever get called? I know it's implementing the
interface. I almost want to say "just have it always return an error,
it should never be called because the client should already have the config."

So adding a comment for me to look into this later.

Signed-off-by: Donny Yung <donaldyung@datawire.io>