Establish correct behavior when returning DNS lookkups
According to RFC 4074, a query for an AAAA record for which no
IPv6 address is found, but an IPv4 is, the return should be
an empty return with no error to indicate that a result exists
when using other query types.

This commit also ensures that the client refrains from calling
the `manager.LookupHost()` multiple times before the first
call has completed. Subsequent calls for the same query are
queued up and waiting for the result of the first one. The
first query is also constrained with a timeout.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>