build-sys: Statically link the CLI
We always did this with `teleproxy`, somehow we lost it along the way
to Telepresence 2.  Bring it back!

Well, we only did it for GOOS=linux with `teleproxy`, and not for
GOOS=darwin because it needed CGo for github.com/datawire/pf.  But we
don't use that anymore now that we have the TUN device, so we can do
it on GOOS=darwin now too!

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>