Limit the DNS requests that result in cluster-side lookups.
This commit adds a set of very common top-level domains for which our
resolver will return not found (or revert to fallback, in case of the
overriding resolver).

The overriding resolver is also tweaked so that it no longer uses the
search path when the last name of a qualified name is equal to a
namespace or "svc".

The max time for the LookupHost is now hard-coded to 4 seconds.

Signed-off-by: Thomas Hallgren <thomas@tada.se>