Fix garbage collection bug
In the last commit where I added RemoveIntercept calls to the
systema-gc, I realized somebody leaving an intercept, deletes the
intercept from the watch, which is then processed by the systema-gc, I
decided to * only * call RemoveIntercept from that loop. Which I think
makes this a little bit simpler.

In my testing, I realized that if there is only one client using
telepresence and they quit before leaving an intercept, the garbage
collection is not able to clean up the intercept (or the domain) from
system a because there is no associated API key.

I think the better longterm solution is to have a System A heartbeat RPC
that informs System A of intercepts that are still "live", but since we
are a little strapped for time right now and that would require a
cross-team effort, I've added a way to keep APIKeys associated wither
intercepts in the state of the manager and only delete them * after *
the garbage collection has informed System A that the intercept + domain
has been removed.

Signed-off-by: Donny Yung <donaldyung@datawire.io>