Move install, update, and delete of traffic-manager to resource package
Moves the code that previously installed, updated, and deleted the
traffic-manager from the connector package to the new resource
package. Much of this is copy/paste but since this is a more generic
mechanism, some parts were rewritten.

One notable change is that the license secret is now always mounted
but optional which should vouch for more consistent behavior when
that secret arrives or is removed (we didn't see that as a reason to
update the traffic-manager, but we should have since it actually added
or removed volumes from the deployment).

Signed-off-by: Thomas Hallgren <thomas@datawire.io>