Get rid of repeated calls to manager.WatchIntercepts()
The client only needs one watcher to keep track of all intercepts. This
commit ensures that the snapshot maintained by that watcher is available
to other code that needs it.

The one and only watcher is now also combined with logic that uses a
`sync.Map` with `chan` values to notify the logic that waits for an
intercept to become active.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>