Configure Agent Image + Registry via config.yml
This commit removes the client-side use of the TELEPRESENCE_AGENT_IMAGE
and TELEPRESENCE_REGISTRY environment variables.  Instead we now
configure these values in the config.yml.  We will likely want to move
the rest of the Environment Variables into the config.yml but this is a
good start (since these are the ones people actually set in practice).

Signed-off-by: Donny Yung <donaldyung@datawire.io>