Lifecycle fix for Webhook Agent Image config
As noted by Thomas, moving these environment variables to the
configuration will make it more apparent how we were using one value for
a lot of things.  With this change, I'm proposing adding an additional
configuration key to separate the AgentImage (a client will use to
specify which image they want to install with their intercept) and
WebhookAgentImage (which is the agent image that the webhook will inject
when it sees the annotation).

This will be easier to document since we can say "if you use
telepresence connect to deploy telepresence, you can set this value to
toggle the image the webhook will inject".

Signed-off-by: Donny Yung <donaldyung@datawire.io>