Ensure that the manager's RoleBinding appoints the correct Role.
The installer created a `RoleBinding` in the Traffic Manager's namespace
that pointed to a non existing role. The `RoleRef` in that binding used
the name "traffic-manager" but the actual `Role` was named
"traffic-manager-<traffic-manager-ns>".

This commit solves the problem by removing "<traffic-manager-ns>" from
the `metadata.name` of both the `Role` and the `RoleRef`. It's redundant
anyway since those objects are namespace scoped.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>