Avoid adding all subcommands except "daemon-foreground" for root daemon.
Attempting to load extensions or provide other options to commands in
the root daemon is pointless. This commit ensures that if the subcommand
is `daemon-foreground`, then only that subcommand is added.

Among other things, this restores the logging functionality in the
daemon which was otherwise initialized too early which cause the config
to be read from the roots config directory rather than the users.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>