Prevent subnet updates when subnets haven't changed.
The `traffic-manager` listens to `Nodes` or `Pods` to determine what
pod-subnets that are used. The `Nodes` or `Pods` may change frequently
although the subnet set remains the same, causing a lot of unnecessary
information to be sent from the `traffic-manager` to connected clients.
Especially o larger clusters.

This commit ensures that changes that don't alter the actual subnets
are ignored and not propagated to the root daemon.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>