Make references in agent's environment work
The `traffic-agent` would receive a copy of the intercepted containers
environment where the names were prefixed with "TEL_APP_" to distinguish
those names from the names found in the agent's own environment. That
was problematic, given that when Kubernetes starts a container, it
interpolates references to other variables in the form "$(NAME)".

This commit removes the prefixes so that the names in the environment
copy are left intact. Instead, it's ensured that the names used in the
agent's own environment are prefixed in such a way that a name collision
is highly unlikely.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>