Fixes an edge case when selecting a container by port number
It's much clearer to first look for a matching container, then if we
don't find one use the first "no ports" container found. (Tho' it's not
clear at all why a no-ports container should be a match, let alone the
first one. Maybe include a comment about why that's the case. I don't
know, so I can't add that.)

#resolves 1926

Signed-off-by: Jeff Williams <jeff.williams@gadgetworks.com>