Use the correct apikey for smoke-tests
Previously we had used an apikey that only exists when you communicate
with Ambassador Cloud to get the Smart Agent version, when you are
manually overriding the Smart Agent image, you don't get that apikey.

I've switched it to use the apikey that we generate when doing an
intercept so this should be more resilient.

Signed-off-by: Donny Yung <donaldyung@datawire.io>