Retry dial when socket exists after ungraceful termination
The daemons will sometimes leave a socket behind in the filesystem when
it is terminated ungracefully (kill -9 or similar). Telepresence will
think that the process is running and try to connect to that socket if
it exists and it will fail because there's no process alive that listens
to it.

Prior to this commit, telepresence would inform the user about the
failed attempt and exit but now, it will instead make an attempt to
remove the socket and if that succeeds, dial again.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>