Fix deadlock in login handler
The access-token "refresh" routine has a select case that calls the
`loginExecutor.getToken` method which in turn would retrieve the `Token`
and potentially end up in a callback writing to the `refreshTimerReset`
channel that was read by another case in that same select, which lead
to a dead-lock because there would be nothing reading that channel.

This commit makes the `refreshTimerReset` channel buffered so that
the deadlock is avoided.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>