Simplify how Tunnel is closed.
A tunnel is closed by one of six things happening at either end (or at
both ends).

1. Read from local connection fails (typically EOF)
2. Write to local connection fails (connection peer closed)
3. Idle timer timed out.
4. Context is cancelled.
5. Disconnect request received from Tunnel peer.
6. DisconnectOK received from Tunnel peer.

When #1 or #2 happens, the Tunnel will send a Disconnect request to
its Tunnel peer, shorten the Idle timer, and then continue to serve
incoming data from the Tunnel peer until a DisconnectOK is received.
Once that happens, it's guaranteed that the Tunnel peer will send no
more messages and the Tunnel is closed.

When #3, #4, or #5 happens, the Tunnel will send a DisconnectOK to its
Tunnel peer and close.

When #6 happens, the Tunnel will simply close.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>