Provide environment variable TELEPRESENCE_INTERCEPT_ID during intercept.
It is sometimes useful to know if a process is running locally on the
workstation, serving an intercept, or if it runs non-intercepted in
a pod, and if the former is true, also know the actual intercept ID
that is being used. The `TELEPRESENCE_INTERCEPT_ID` environment variable
can help with both cases. It is only set for processes running locally
on the workstation.

Having the intercept ID available can be helpful when dealing with
pub/sub systems like Kafka, who can then have all processes that doesn't
have the `TELEPRESENCE_INTERCEPT_ID` set, filter out all messages that
contain an `x-intercept-id` header and those who have the environment
variable set, filter based on the matching `x-intercept-id` header. This
will assure that such messages only are consumed by the correct process.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>