Let the telepresence status command report correct DNS IP
The `telepresence status` would report that both local and remote
IP is "nil" on most platforms. The local IP is only used on Linux and
only by the overriding DNS resolver and should only be reported when
it's set. The remote IP is always set (but perhaps not in the config)
and the current value should be reported.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>